requirements.txt
# Core Backend
fastapi
uvicorn

# LLM + RAG Stack
langchain
openai
pinecone-client

# Embeddings & NLP
sentence-transformers
tiktoken

# Text Chunking & Parsing
pdfplumber
python-docx
PyMuPDF

# Frontend
streamlit

# Evaluation
nltk
ragas==0.0.10  # optional, if using RAGAS metrics

# Utilities
python-dotenv
pydantic

# Testing
pytest

# Formatting & Linting
black
flake8
